+++ vueManager/src/entries/extra-fields.js
@@ -9,6 +9,7 @@ import { createApp } from 'vue';
 import { createPinia } from 'pinia';
 import PrimeVue from 'primevue/config';
 import Aura from '@primevue/themes/aura';
+import { getPrimeVueLocale } from "@vuetools/usePrimeVueLocale";
 import 'primeicons/primeicons.css';
 
 import ConfirmationService from 'primevue/confirmationservice';
@@ -30,9 +31,10 @@ function createVueApp() {
     theme: {
       preset: Aura,
       options: {
-        darkModeSelector: 'none'
-      }
-    }
+        darkModeSelector: "none",
+      },
+    },
+    locale: getPrimeVueLocale(),
   });
 
   app.use(ConfirmationService);