+++ vueManager/src/entries/import.js
@@ -8,6 +8,7 @@ import '../scss/primevue.scss';
 import { createApp } from 'vue';
 import PrimeVue from 'primevue/config';
 import Aura from '@primevue/themes/aura';
+import { getPrimeVueLocale } from "@vuetools/usePrimeVueLocale";
 import 'primeicons/primeicons.css';
 
 import ConfirmationService from 'primevue/confirmationservice';
@@ -27,11 +28,12 @@ function createVueApp() {
     theme: {
       preset: Aura,
       options: {
-        prefix: 'p',
-        darkModeSelector: '.ms3-dark-mode',
-        cssLayer: false
-      }
-    }
+        prefix: "p",
+        darkModeSelector: ".ms3-dark-mode",
+        cssLayer: false,
+      },
+    },
+    locale: getPrimeVueLocale(),
   });
 
   vueApp.use(ConfirmationService);