+++ vueManager/src/entries/links.js
@@ -9,6 +9,7 @@ import PrimeVue from 'primevue/config'
 import Aura from '@primeuix/themes/aura'
 import ToastService from 'primevue/toastservice'
 import ConfirmationService from 'primevue/confirmationservice'
+import { getPrimeVueLocale } from "@vuetools/usePrimeVueLocale";
 import LinksGrid from '../components/LinksGrid.vue'
 
 // Mount point ID
@@ -38,11 +39,12 @@ function mountApp() {
     theme: {
       preset: Aura,
       options: {
-        darkModeSelector: '.dark-mode',
-        cssLayer: false
-      }
-    }
-  })
+        darkModeSelector: ".dark-mode",
+        cssLayer: false,
+      },
+    },
+    locale: getPrimeVueLocale(),
+  });
 
   app.use(ToastService)
   app.use(ConfirmationService)