+++ vueManager/src/entries/order.js
@@ -9,6 +9,7 @@ import { createApp } from 'vue';
 import { createPinia } from 'pinia';
 import PrimeVue from 'primevue/config';
 import Aura from '@primevue/themes/aura';
+import { getPrimeVueLocale } from "@vuetools/usePrimeVueLocale";
 import 'primeicons/primeicons.css';
 
 import ConfirmationService from 'primevue/confirmationservice';
@@ -21,7 +22,6 @@ import OrderView from '../components/OrderView.vue';
  */
 function createVueApp() {
   const app = createApp(OrderView);
-
   const pinia = createPinia();
   app.use(pinia);
 
@@ -29,9 +29,10 @@ function createVueApp() {
     theme: {
       preset: Aura,
       options: {
-        darkModeSelector: 'none'
-      }
-    }
+        darkModeSelector: "none",
+      },
+    },
+    locale: getPrimeVueLocale(),
   });
 
   app.use(ConfirmationService);