+++ vueManager/src/entries/payments.js
@@ -8,6 +8,7 @@ import '../scss/primevue.scss'
 import { createApp } from 'vue'
 import PrimeVue from 'primevue/config'
 import Aura from '@primeuix/themes/aura'
+import { getPrimeVueLocale } from "@vuetools/usePrimeVueLocale";
 import 'primeicons/primeicons.css'
 
 import ConfirmationService from 'primevue/confirmationservice'
@@ -19,16 +20,16 @@ import PaymentsGrid from '../components/PaymentsGrid.vue'
  * Creates and configures Vue application
  */
 function createVueApp() {
-  const app = createApp(PaymentsGrid)
-
+  const app = createApp(PaymentsGrid);
   app.use(PrimeVue, {
     theme: {
       preset: Aura,
       options: {
-        darkModeSelector: 'none'
-      }
-    }
-  })
+        darkModeSelector: "none",
+      },
+    },
+    locale: getPrimeVueLocale(),
+  });
 
   app.use(ConfirmationService)
   app.use(ToastService)