+++ vueManager/src/entries/utilities-gallery.js
@@ -8,6 +8,7 @@ import '../scss/primevue.scss';
 import { createApp } from 'vue';
 import PrimeVue from 'primevue/config';
 import Aura from '@primevue/themes/aura';
+import { getPrimeVueLocale } from "@vuetools/usePrimeVueLocale";
 import 'primeicons/primeicons.css';
 
 import ToastService from 'primevue/toastservice';
@@ -21,16 +22,16 @@ let app = null;
  */
 function createVueApp(props = {}) {
   const vueApp = createApp(UtilitiesGallery, props);
-
   vueApp.use(PrimeVue, {
     theme: {
       preset: Aura,
       options: {
-        prefix: 'p',
-        darkModeSelector: '.ms3-dark-mode',
-        cssLayer: false
-      }
-    }
+        prefix: "p",
+        darkModeSelector: ".ms3-dark-mode",
+        cssLayer: false,
+      },
+    },
+    locale: getPrimeVueLocale(),
   });
 
   vueApp.use(ToastService);